Cambios para empezar a validar las reglas

+++ trunk/grails-app/domain/user/Reputation.groovy
@@ -6,7 +6,7 @@ class Reputation {
 	
 	UserRole userRole
 	static belongsTo	= [user: User]
-	double value
+	long value
 	
 	
     static mapping = {
